+++ fitstable.c
@@ -130,7 +130,7 @@ static int class_check;
 
 /* Pointers to parent class methods which are extended by this class. */
 static int (* parent_equal)( AstObject *, AstObject *, int * );
-static int (* parent_getobjsize)( AstObject *, int * );
+static size_t (* parent_getobjsize)( AstObject *, int * );
 static void (* parent_addcolumn)( AstTable *, const char *, int, int, int *, const char *, int * );
 
 #if defined(THREAD_SAFE)
@@ -177,7 +177,7 @@ static AstFitsChan *GetTableHeader( AstFitsTable *, int * );
 static char *MakeKey( const char *, int, char *, int, int * );
 static int ColumnNull( AstFitsTable *, const char *, int, int, int *, int *, int * );
 static int Equal( AstObject *, AstObject *, int * );
-static int GetObjSize( AstObject *, int * );
+static size_t GetObjSize( AstObject *, int * );
 static size_t ColumnSize( AstFitsTable *, const char *, int * );
 static void AddColumn( AstTable *, const char *, int, int, int *, const char *, int * );
 static void Copy( const AstObject *, AstObject *, int * );
@@ -1415,7 +1415,7 @@ f     AST_COLUMNNULL functiom.
    *nelem = nel*nrow;
 }
 
-static int GetObjSize( AstObject *this_object, int *status ) {
+static size_t GetObjSize( AstObject *this_object, int *status ) {
 /*
 *  Name:
 *     GetObjSize
@@ -1428,7 +1428,7 @@ static int GetObjSize( AstObject *this_object, int *status ) {
 
 *  Synopsis:
 *     #include "fitstable.h"
-*     int GetObjSize( AstObject *this, int *status )
+*     size_t GetObjSize( AstObject *this, int *status )
 
 *  Class Membership:
 *     FitsTable member function (over-rides the astGetObjSize protected
@@ -1454,7 +1454,7 @@ static int GetObjSize( AstObject *this_object, int *status ) {
 
 /* Local Variables: */
    AstFitsTable *this;            /* Pointer to FitsTable structure */
-   int result;                /* Result value to return */
+   size_t result;             /* Result value to return */
 
 /* Initialise. */
    result = 0;
