+++ pointlist.c
@@ -135,7 +135,7 @@ static int class_check;
 static AstMapping *(* parent_simplify)( AstMapping *, int * );
 static AstPointSet *(* parent_transform)( AstMapping *, AstPointSet *, int, AstPointSet *, int * );
 static const char *(* parent_getattrib)( AstObject *, const char *, int * );
-static int (* parent_getobjsize)( AstObject *, int * );
+static size_t (* parent_getobjsize)( AstObject *, int * );
 static int (* parent_testattrib)( AstObject *, const char *, int * );
 static void (* parent_clearattrib)( AstObject *, const char *, int * );
 static void (* parent_setattrib)( AstObject *, const char *, int * );
@@ -199,7 +199,7 @@ static AstPointSet *Transform( AstMapping *, AstPointSet *, int, AstPointSet *, 
 static AstRegion *RegBasePick( AstRegion *, int, const int *, int * );
 static int GetClosed( AstRegion *, int * );
 static int GetListSize( AstPointList *, int * );
-static int GetObjSize( AstObject *, int * );
+static size_t GetObjSize( AstObject *, int * );
 static int RegPins( AstRegion *, AstPointSet *, AstRegion *, int **, int * );
 static void Copy( const AstObject *, AstObject *, int * );
 static void PointListPoints( AstPointList *, AstPointSet **, int *);
@@ -463,7 +463,7 @@ static int GetListSize( AstPointList *this, int *status ) {
    return astGetNpoint( ((AstRegion *) this)->points );
 }
 
-static int GetObjSize( AstObject *this_object, int *status ) {
+static size_t GetObjSize( AstObject *this_object, int *status ) {
 /*
 *  Name:
 *     GetObjSize
@@ -476,7 +476,7 @@ static int GetObjSize( AstObject *this_object, int *status ) {
 
 *  Synopsis:
 *     #include "pointlist.h"
-*     int GetObjSize( AstObject *this, int *status )
+*     size_t GetObjSize( AstObject *this, int *status )
 
 *  Class Membership:
 *     PointList member function (over-rides the astGetObjSize protected
@@ -502,7 +502,7 @@ static int GetObjSize( AstObject *this_object, int *status ) {
 
 /* Local Variables: */
    AstPointList *this;         /* Pointer to PointList structure */
-   int result;                /* Result value to return */
+   size_t result;             /* Result value to return */
 
 /* Initialise. */
    result = 0;
